:truck: Split and move class
* namespace 的に適切なディレクトリに移動させる
* Extension とメイン実装のクラスを分ける
