+++ Assets/Scripts/UniRx/UnityEngineBridge/Triggers/KeyEventComponentTrigger.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 145eacf85f8c4328be11e55e69a2ef3e
+timeCreated: 1518624405
